+++ plugin/api-gradle-plugin/src/main/java/com/blankj/api/Config.groovy
@@ -14,6 +14,4 @@ class Config {
     ]
 
     public static final String FILE_SEP = System.getProperty("file.separator")
-
-    public static final String API_UTILS_CLASS = 'com.blankj.utilcode.util.ApiUtils'.replace('.', FILE_SEP) + '.class'
 }