+++ plugin/api-gradle-plugin/src/test/java/com/blankj/api/ApiUtils.java
@@ -1,4 +1,4 @@
-package com.blankj.utilcode.util;
+package com.blankj.api;
 
 import com.android.annotations.NonNull;