+++ plugin/bus-gradle-plugin/src/main/java/com/blankj/bus/BusInject.groovy
@@ -1,6 +1,5 @@
 package com.blankj.bus
 
-import com.blankj.bus.util.LogUtils
 import com.blankj.bus.util.ZipUtils
 import org.apache.commons.io.FileUtils
 import org.objectweb.asm.ClassReader
@@ -10,19 +9,18 @@ import org.objectweb.asm.ClassWriter
 class BusInject {
 
     static void start(Map<String, BusInfo> busMap, File busUtilsTransformFile, String busUtilsClass) {
-        LogUtils.l("===>" + busUtilsTransformFile)
         if (busUtilsTransformFile.getPath().endsWith(".jar")) {
             String jarPath = busUtilsTransformFile.getAbsolutePath()
             String decompressedJarPath = jarPath.substring(0, jarPath.length() - 4);
             File decompressedJar = new File(decompressedJarPath)
             ZipUtils.unzipFile(busUtilsTransformFile, decompressedJar)
 
-            File apiUtilsFile = new File(
+            File busUtilsFile = new File(
                     decompressedJarPath + Config.FILE_SEP +
                             busUtilsClass.replace('.', Config.FILE_SEP) + '.class'
             )
 
-            inject2BusUtils(apiUtilsFile, busMap, busUtilsClass)
+            inject2BusUtils(busUtilsFile, busMap, busUtilsClass)
 
             FileUtils.forceDelete(busUtilsTransformFile)
             ZipUtils.zipFiles(Arrays.asList(decompressedJar.listFiles()), busUtilsTransformFile)
@@ -44,5 +42,4 @@ class BusInject {
         cr.accept(cv, ClassReader.SKIP_FRAMES);
         FileUtils.writeByteArrayToFile(apiUtilsFile, cw.toByteArray())
     }
-
 }